feat: Course totalDuration 필드 타입 변경 (int -> double)

+++ backend/domain-modules/mountain-domain/src/main/java/com/softeer/domain/Course.java
@@ -6,7 +6,7 @@ public record Course(
         long id,
         String name,
         double totalDistance,
-        int totalDuration,
+        double totalDuration,
         Level level,
         String imageUrl
 ) {
